requirements.txt
alpaca-trade-api==3.0.1
python-dotenv==1.0.1
pandas==2.2.1
requests==2.31.0
schedule==1.2.1
